Refactor: 부족한 PropTypes 기술 및 디렉토리 정리
디렉토리에 있는 파일들을 일관성있게 정리하고 PropTypes가 빠진부분을
보충함
